Make code style more consistent with rest of the project
