+++ src/parsers/contest/HKOIOnlineJudgeContestParser.ts
@@ -1,9 +1,9 @@
-import { HKoiJudgeProblemParser } from '../problem/HKoiJudgeProblemParser';
+import { HKOIOnlineJudgeProblemParser } from '../problem/HKOIOnlineJudgeProblemParser';
 import { SimpleContestParser } from '../SimpleContestParser';
 
-export class HKoiJudgeContestParser extends SimpleContestParser {
+export class HKOIOnlineJudgeContestParser extends SimpleContestParser {
   protected linkSelector = 'table.table-hover tbody tr td:nth-child(2) a';
-  protected problemParser = new HKoiJudgeProblemParser();
+  protected problemParser = new HKOIOnlineJudgeProblemParser();
 
   public getMatchPatterns(): string[] {
     return ['https://judge.hkoi.org/contest/*'];
